Perform a huge rename from "payload" to "action data"
I had been using "payload" to refer to the object that gets put inside a card action for it to be sent back to the bot when the user clicks the button, when it then becomes an "incoming payload." I decided this was ambiguous because "payload" normally refers to the entire body/content of an HTTP request, and so the real payloads exchanged between a bot and a channel always contain more than just what I've been calling a payload. The word payload is also used as a property of the channel data in an activity that's formed from an interactive message from Slack, and it contains what looks like the whole body of the request the Slack connector received from Slack rather than just the value of the card action

The term "action data" is much better. I've shortened it to just "data" in most cases so that it's even shorter than "payload," and that term is already used in Adaptive Cards. It's intuitively clear that "action data" and "data" refer to the same thing when seeing them used interchangeably