Updated [Configure your Actions SDK project]
It looks like this readme was originally written with the project in Isolation instead of being consolidated into a much larger Repository. Since it now is part of a HUGE repo its not perhaps as self evident that the user is actually expected to download the whole thing and then find the action folder that is in the same directory as this readme in the resulting code base.

At least it wasn't obvious to me and cost me some time. Update hopefully will help the next guy.