Change to UTF8 encoding in the body validation when requesting authorization (#439)
* StreamReader overload to specify UTF8 encoding type

StreamReader overload to UTF8 encoding type in order to have more compatibility with the character sets of other languages ​​apart from English.  
since this can generate incompatibility in certain cases such as in #435

* Change from ASCII to UTF8 encoding in _requestBodyByte

Change from ASCII to UTF8 encoding type in the body request to have more compatibility with the character sets of other languages ​​apart from English.  
since this can generate incompatibility in certain cases such as in #435